Name lookup (cache phone numbers, change phone book ready() checks).
Cache phone numbers (as longs) for each index in the SIM card phone book. Then, when a call comes, find the index (if any) with matching number and look up the name.

Change the PhoneBook class to remove the internal "while (!ready());" checks. Instead, these are done in the sketch to allow various timeout strategies.